backend/requirements.txt
fastapi
uvicorn
anthropic
apify-client
python-dotenv